+++ cargo-tauri.ps1
@@ -1,5 +1,0 @@
-# Wrapper script to use npm tauri instead of cargo tauri
-$projectRoot = "D:\Users\feiqin\01code\99study\httpulse"
-Set-Location $projectRoot
-npm run tauri @args
-