+++ src/constants/common.ts
@@ -1,1 +1,8 @@
-export const appName = "Cyber API";
+/*
+ * @Author: feiqin
+ * @Date: 2025-10-23 14:14:54
+ * @LastEditors: feiqin
+ * @LastEditTime: 2025-10-23 15:44:50
+ * @Description: 
+ */
+export const appName = "HTTPulse";